back/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

constant==0.0.4
jsons==1.6.3
Pillow==9.4.0
PyAutoGUI==0.9.53
python-dotenv==0.21.1
requests==2.28.2
